requirements.txt
autopep8
datasets
evaluate
huggingface-hub
matplotlib
numpy
pandas
python-dotenv
requests
scikit-learn
scipy
tokenizers
torch
torchaudio
torchvision
transformers
conllu
seaborn 



